feat: 리뷰 작성 기능 구현
- 별점 업데이트
- 리뷰 유효성 검증

+++ src/main/java/com/hyundai/app/store/dto/StoreResDto.java
@@ -24,9 +24,9 @@ public class StoreResDto {
     private String description;
     private int phone;
     private int floor;
-    private int avgScore;
+    private float avgScore;
 
-    private List<Hashtag> popularHashtags; // 사실 이름만 필요함!
+    private List<Hashtag> popularHashtags;
     private List<ReviewResDto> reviews;
 
     public static StoreResDto of(Store store) {
